fix: Move filtered Constants to hawkeye-core
