+++ hawkeye-command/pom.xml
@@ -54,13 +54,4 @@
             <artifactId>slf4j-simple</artifactId>
         </dependency>
     </dependencies>
-
-    <build>
-        <resources>
-            <resource>
-                <directory>src/main/resources</directory>
-                <filtering>true</filtering>
-            </resource>
-        </resources>
-    </build>
 </project>
