+++ hawkeye-command/src/main/java/io/korandoru/hawkeye/command/HawkEyeVersionProvider.java
@@ -16,11 +16,12 @@
 
 package io.korandoru.hawkeye.command;
 
+import io.korandoru.hawkeye.core.Constants;
 import picocli.CommandLine;
 
 public class HawkEyeVersionProvider implements CommandLine.IVersionProvider {
     @Override
     public String[] getVersion() {
-        return new String[] { CommandConstants.VERSION };
+        return new String[] { Constants.VERSION };
     }
 }
