+++ hawkeye-core/src/main/resources/META-INF/native-image/io.korandoru.hawkeye/hawkeye-core/resource-config.json
@@ -3,6 +3,9 @@
     "includes": [
       {
         "pattern": ".*\\.txt"
+      },
+      {
+        "pattern": "hawkeye.properties"
       }
     ]
   }
